Initial post
This script has been banging around in the bottom of my toolbox for a while, it is time to share it. It has been lightly cleansed and tidied up so it is freestanding. There are a number of prereqs commented inline code. 

TODO: The prereqs etc need to move to README.md